add backups files to all starr apps
