+++ lidarr/lidarr.go
@@ -359,3 +359,15 @@ func (l *Lidarr) GetHistory(maxRecords int) (*History, error) {
 
 	return &history, nil
 }
+
+// GetBackupFiles returns all available Lidarr backup files.
+// Use GetBody to download a file using BackupFile.Path.
+func (l *Lidarr) GetBackupFiles() ([]*starr.BackupFile, error) {
+	var output []*starr.BackupFile
+
+	if err := l.GetInto("v1/system/backup", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(system/backup): %w", err)
+	}
+
+	return output, nil
+}
