+++ readarr/readarr.go
@@ -304,3 +304,15 @@ func (r *Readarr) GetHistory(maxRecords int) (*History, error) {
 
 	return &history, nil
 }
+
+// GetBackupFiles returns all available Readarr backup files.
+// Use GetBody to download a file using BackupFile.Path.
+func (r *Readarr) GetBackupFiles() ([]*starr.BackupFile, error) {
+	var output []*starr.BackupFile
+
+	if err := r.GetInto("v1/system/backup", nil, &output); err != nil {
+		return nil, fmt.Errorf("api.Get(system/backup): %w", err)
+	}
+
+	return output, nil
+}
